983. Minimum Cost For Tickets
